refactor(desktop): extract useShellSearch from app-shell
Step 3c follow-up (seam 4) of the #1043 plan. The search-modal slice -
the open flag, the funnel-bridge initial query, the scroll-target anchor,
the close handler (restores focus to the sidebar trigger), and the stable
search-thread dep + navigate callback - moves into a useShellSearch hook.

AppShell passes its openSessionInChatRef so the navigate callback stays
memoized while always calling the latest opener; the palette's
onOpenSearchModal keeps using the returned setters. Isomorphic to the
existing shell-slice hooks.

Pure code motion - no behavior change. use-shell-search.ts is registered
in the source helper; the search-modal-lifecycle and modal-lifecycle
contracts already use combined source (the SearchModal mount lives in
AppShellOverlays) so they cover the new file without changes.
